S'argo Data Repository v2
This will be the new data repository for v2+. When v2 of S'argo is released, the original data repository will no longer be maintained.

So, why the new data repository? Here are some new features planned that I can list from the top of my head:
- All character/weapon images share the same folder; no more individual banner specific folders
- All character images now have the character name after their prefix; this is so that characters with duplicate prefixes can exist without the (#) i.e: "Nightwear" Asuna/Leafa/Silica/Lisbeth/Sinon/Yuuki and "Promise at that Time" Asuna/Kirito.
- New System folder: All system related images (scout backgrounds, GUI thumbnails, etc) can be found in this folder rather than different folders in /images/.
Anyways, that's basically the reason why I need to make a new data repository. Overwriting the old repository will cause issues for users who still plan to use the v1 S'argo so i decided to just keep it there.